feat: Non-renderframe-synced keysound/updates
`timer.update()` is tied to `MainController::render` call, and it is
used as a source of truth for keysound & judging system. This patch
creates a separate thread that continuously updates the timer.

Note that this patch isn't yet tested on the wild or had undergone any
unit tests, so it might break some other functionalities. Further
testing required.